Solved some bugs and implemented OutputModel for Transaction.

+++ server/src/main/kotlin/pt/isel/ps/LI61N/g30/server/model/domain/Toll.kt
@@ -49,7 +49,7 @@ data class Toll(
         val geolocation_longitude: Float,
 
         @Column(name="azimuth", unique = true, nullable = true)
-        val azimuth: Float?,
+        val azimuth: Double?,
 
         @Column(name="region", unique = true, nullable = true)
         val region: String?,
